Fixed bug with Logging, AGAIN!

+++ database/migrations/2026_03_27_182929_drop_updated_at_from_logs_table.php
@@ -8,9 +8,11 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('logs', function (Blueprint $table) {
-            $table->dropColumn('updated_at');
-        });
+        if (Schema::hasColumn('logs', 'updated_at')) {
+            Schema::table('logs', function (Blueprint $table) {
+                $table->dropColumn('updated_at');
+            });
+        }
     }
 
     public function down(): void
